changed the type of value attribute

+++ src/Measurement.java
@@ -2,7 +2,7 @@ public class Measurement {
     private final double value;
     private final Unit unit;
 
-    public Measurement(int value, Unit unit) {
+    public Measurement(double value, Unit unit) {
         this.value = value;
         this.unit = unit;
     }
